updates timeline to allow looping back to earlier Views (closes #145)
adding a new method to the api/useTimeline that lets you jump to a specific route.  this has behavior of allowing “forced” navigations.  this simplifies serveral other parts of the code including the routejumper.vue in dev mode because we just treat any programmatic navigation as allowed but don’t allow browser notifications.  the ideas is that any page request coming from the code will have been wanted.  simplified!