feat: prevent issues with changing the dimensionality of a nested table
in a nested table if you resize the number of rows after the nested table is created it can lead to undesired behavior.  instead all methods that  can do this raise a warning to prevent it.  it is best practice to call the head(), tail(), sample() methods before setting up the nested tables
- docs and tests updated