feat: serialize vue components
when a data field include a vue component (useful for dynamic component rendering) the system needs to properly serialize it without rendering the entire vue component SFC into the localstorage.  instead it stores a name for the component and updates it as it goes.   the system scanes for imported/referenced vue components and makes sure they are availble for importing.
- includes some tests